move assets into docs
