+++ src/docs/components/variants/image/ImageStyle1to1.tsx
@@ -1,6 +1,6 @@
 import Example from '../../../shared/components/Example';
 import Image from '../../../../library/components/Image';
-import image1to1 from '../../../../assets/images/1to1.jpg';
+import image1to1 from '../../../../docs/assets/images/1to1.jpg';
 
 const ImageStyle1to1 = (): JSX.Element => {
   return (
