+++ src/docs/components/variants/image/WithCaption.tsx
@@ -1,6 +1,6 @@
 import Example from '../../../shared/components/Example';
 import Image from '../../../../library/components/Image';
-import image16to9 from '../../../../assets/images/16to9.jpg';
+import image16to9 from '../../../../docs/assets/images/16to9.jpg';
 
 const WithCaption = (): JSX.Element => {
   return (
