+++ src/docs/components/variants/modals/LogoModal.tsx
@@ -2,7 +2,7 @@ import { useRef } from 'react';
 import Button from '../../../../library/components/Button';
 import Modal, { ModalRef } from '../../../../library/components/Modal';
 import Example from '../../../shared/components/Example';
-import BSCLogo from '../../../../assets/images/logo-bsc-tagline.svg';
+import BSCLogo from '../../../../docs/assets/images/logo-bsc-tagline.svg';
 import Link from '../../../../library/components/Link';
 
 const LogoModal = (): JSX.Element => {
