+++ src/docs/components/variants/navPrimary/SimpleNavPrimary.tsx
@@ -1,6 +1,6 @@
 import NavPrimary from '../../../../library/components/navigation/navPrimary/NavPrimary';
 
-import logoBSC from '../../../../assets/images/logo-bsc-tagline.svg';
+import logoBSC from '../../../../docs/assets/images/logo-bsc-tagline.svg';
 import { simpleData, utilityData } from './navPrimaryData';
 import Example from '../../../shared/components/Example';
